tools/format-converter/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "format-converter"
version = "1.0.0"
description = "Convert between academic document formats with citation preservation"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Deep Biblio Tools Team", email = "team@deepbiblio.tools"}
]
keywords = ["converter", "latex", "markdown", "bibtex", "academic", "citations"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing :: Markup",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0",
    "pypandoc>=1.11",
    "bibtexparser>=1.4",
    "markdown>=3.4",
    "python-docx>=0.8",
    "rich>=13.0",
    "PyYAML>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
format-converter = "format_converter.cli:main"

[project.urls]
Homepage = "https://github.com/deep-biblio-tools/deep-biblio-tools"
Documentation = "https://deep-biblio-tools.readthedocs.io/tools/format-converter"
"Bug Tracker" = "https://github.com/deep-biblio-tools/deep-biblio-tools/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/format_converter"]
